feat: Add new image to the 'images' folder
This commit introduces a new image that has been added to the 'images' folder. The image file enhances the visual content of the project and can be utilized for various purposes such as documentation, user interfaces, or any other relevant context. This addition aligns with the ongoing development and enrichment of the project's assets.